grid_bot/requirements.txt
pybit>=5.6.0
python-dotenv
requests
